+++ trunk/Watchdogd/Syn_cadran.c
@@ -83,7 +83,7 @@
 
     g_snprintf( requete, sizeof(requete),                                                                      /* Requete SQL */
                 "INSERT INTO %s(syn_id,type,bitctrl,posx,posy,angle,tech_id,acronyme,fleche_left,nb_decimal)"
-                " VALUES (%d,%d,%d,%d,%d,'%f','%s','%s','%d','%d')", NOM_TABLE_CADRAN,
+                " VALUES (%d,%d,%d,%d,%d,'%d','%s','%s','%d','%d')", NOM_TABLE_CADRAN,
                 cadran->syn_id, cadran->type, cadran->bit_controle,
                 cadran->position_x, cadran->position_y, cadran->angle, tech_id, acronyme, cadran->fleche_left, cadran->nb_decimal );
     g_free(tech_id);
@@ -161,7 +161,7 @@
        cadran->bit_controle = atoi(db->row[3]);                                                                 /* Ixxx, Cxxx */
        cadran->position_x   = atoi(db->row[4]);                                                  /* en abscisses et ordonnées */
        cadran->position_y   = atoi(db->row[5]);
-       cadran->angle        = atof(db->row[6]);
+       cadran->angle        = atoi(db->row[6]);
        g_snprintf( cadran->tech_id,  sizeof(cadran->tech_id),  "%s" ,db->row[7] );               /* Recopie dans la structure */
        g_snprintf( cadran->acronyme, sizeof(cadran->acronyme), "%s" ,db->row[8] );               /* Recopie dans la structure */
        cadran->fleche_left  = atof(db->row[9]);
@@ -228,7 +228,7 @@
 
     g_snprintf( requete, sizeof(requete),                                                                      /* Requete SQL */
                 "UPDATE %s SET "
-                "type=%d,bitctrl=%d,posx=%d,posy=%d,angle='%f',tech_id='%s',acronyme='%s',fleche_left='%d',nb_decimal='%d'"
+                "type=%d,bitctrl=%d,posx=%d,posy=%d,angle='%d',tech_id='%s',acronyme='%s',fleche_left='%d',nb_decimal='%d'"
                 " WHERE id=%d;", NOM_TABLE_CADRAN,
                 cadran->type, cadran->bit_controle,
                 cadran->position_x, cadran->position_y, cadran->angle,
